getting cleaner, deciding where overlay banner carousel state goes
