+++ react-client/src/components/overlay/OverlayCarousel.jsx
@@ -3,7 +3,6 @@ import React from 'react';
 import {
   CarouselSection,
   CarouselWrapper,
-  CarouselImage,
   CarouselImageWrapper,
 } from '../styled-components/overlay/OverlayStyles.jsx';
 
@@ -12,19 +11,7 @@ export default props => {
     <CarouselWrapper>
       <CarouselSection>
         <CarouselImageWrapper>
-          {props.render()}
-          {/* {props.imgs.map((item, index, array) => {
-            return (
-              <CarouselImage
-                key={index}
-                src={item}
-                onClick={() => {
-                  props.handleImageClick(index);
-                }}
-                index={index}
-              />
-            );
-          })} */}
+          {props.render(props)}
         </CarouselImageWrapper>
       </CarouselSection>
     </CarouselWrapper>
