+++ react-client/src/components/productGallery/NotFound.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import NotFound from '../styled-components/NotFound.jsx';
+
+export default props => {
+  return (
+    <NotFound fourOhFour={props.fourOhFour}>
+      Not Found - Please Try Another Product
+    </NotFound>
+  );
+};
