+++ react-client/src/components/productGallery/ProductBanner.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import Image from './Image.jsx';
+
+export default props => {
+  return <Image src={props.bannerImg} />;
+};
