+++ react-client/src/components/productGallery/ProductCarousel.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { CarouselImage } from '../styled-components/overlay/OverlayStyles.jsx';
+
+export default props => {
+  return props.carouselImgs.map((item, index, collection) => {
+    return <CarouselImage key={index} src={item} index={index} />;
+  });
+};
