+++ react-client/src/components/productGallery/ProductImage.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Image from './Image.jsx';
+import MagnifyingGlass from './MagnifyingGlass.jsx';
+
+
+export default props => {
+  return (
+    <Image
+      src={props.bannerImg}
+      scrollOverDisplay={true}
+      renderScrollOverDisplay={() => <MagnifyingGlass />}
+    />
+  );
+};
