+++ react-client/src/components/productGallery/index.jsx
@@ -1,0 +1,79 @@
+/* fetches product data and renders product gallery */
+
+import React from 'react';
+import axios from 'axios';
+
+import NotFound from './NotFound.jsx';
+
+import OverlayRenderer from '../overlay/OverlayRenderer.jsx';
+
+import ProductImage from './ProductImage.jsx';
+import ProductBanner from './ProductBanner.jsx';
+import ProductCarousel from './ProductCarousel.jsx';
+
+class ProductGallery extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      bannerImg: '',
+      carouselImgs: [],
+      fourOhFour: false,
+    };
+  }
+
+  componentDidMount() {
+    let url = new URL(window.location.href);
+    let productName = 'test1';
+    if (url.pathname !== '/') {
+      productName = url.pathname.split('/')[2];
+    }
+
+    axios
+      .get(`/productImages/${productName}`)
+      .then(res => {
+        this.setState({
+          bannerImg: res.data.bannerImageUrl,
+          carouselImgs: res.data.images.split(','),
+          fourOhFour: false,
+        });
+      })
+      .catch(err => {
+        if (err.message === 'Request failed with status code 404') {
+          this.setState({
+            fourOhFour: true,
+          });
+        } else {
+          console.error(error);
+        }
+      });
+  }
+
+  render() {
+    return (
+      <div>
+        <NotFound fourOhFour={this.state.fourOhFour}/>
+
+        <OverlayFragment
+          // render what triggers the overlay
+          render={() => {
+            return (
+              <ProductImage
+                scrollOverDisplay={this.state.scrollOverDisplay}
+                bannerImg={this.state.bannerImg}
+              />
+            );
+          }}
+          // render what is in the overlay
+          renderBanner={() => (
+            <ProductBanner bannerImg={this.state.bannerImg} />
+          )}
+          renderCarousel={() => (
+            <ProductCarousel carouselImgs={this.state.carouselImgs} />
+          )}
+        />
+      </div>
+    );
+  }
+}
+
+export default ProductGallery;
