--- react-client/src/config.js
@@ -1,3 +1,0 @@
-export default {
-  X_BUTTON_URL: 'https://cdn3.iconfinder.com/data/icons/iconic-1/32/x_alt-512.png'
-}
